feat(dactle): abstract db class
