+++ packages/dactle/package.json
@@ -1,7 +1,7 @@
 {
   "private": false,
   "name": "@darwinia/dactle",
-  "version": "0.1.5",
+  "version": "0.1.6",
   "description": "darwinia ctl",
   "homepage": "https://github.com/darwinia-network/darwinia.js",
   "repository": {
@@ -18,9 +18,9 @@
   "dependencies": {
     "@darwinia/api": "0.1.26",
     "@darwinia/util": "0.1.26",
+    "ioredis": "^4.16.3",
     "js-yaml": "^3.13.1",
     "node-telegram-bot-api": "^0.40.0",
-    "redis": "^3.0.2",
     "yargs": "^15.3.1"
   },
   "scripts": {
