+++ packages/dactle/src/_db.ts
@@ -1,103 +1,12 @@
-import fs from "fs"
-import path from "path"
+export default abstract class BotDb {
+    constructor() { }
 
-/**
- * Database of bot
- */
-interface IBotScheme {
-    addrs: string[],
-    users: Record<string, number>,
-    supply: Record<string, number>,
+    public abstract addAddr(addr: string): void;
+    public abstract hasReceived(addr: string): boolean;
+
+    public abstract nextDrop(id: number, interval: number): number;
+    public abstract lastDrop(id: number, last: number): void;
+
+    public abstract hasSupply(date: string, supply: number): boolean;
+    public abstract burnSupply(date: string, supply: number): void;
 }
-
-/**
- * This database is just for the telegram bot currently now
- */
-export class BotDb {
-    private _: IBotScheme;
-    private path: string;
-
-    constructor(
-        _path: string,
-        supply = 0,
-    ) {
-        // ensure the path of database
-        if (!fs.existsSync(_path)) {
-            fs.mkdirSync(path.dirname(_path), { recursive: true });
-            fs.writeFileSync(_path, "");
-            this._ = {
-                addrs: [],
-                users: {},
-                supply: {
-                    [new Date().toJSON().slice(0, 10)]: supply,
-                },
-            };
-        } else {
-            this._ = JSON.parse(fs.readFileSync(_path).toString());
-        }
-
-        this.path = _path;
-    }
-
-    /**
-     * Validate address
-     */
-    public addAddr(addr: string) {
-        this._.addrs.push(addr);
-        this.save();
-    }
-
-    public hasReceived(addr: string) {
-        if (this._.addrs.indexOf(addr) > -1) {
-            return true;
-        }
-        return false;
-    }
-
-    /**
-     * Validate user id
-     */
-    public nextDrop(id: string, interval: number): number {
-        const last: number = this._.users[id];
-        if (last !== undefined) {
-            const sub = interval - ((new Date().getTime() - last) / 1000 / 60 / 60);
-            return Math.floor(sub);
-        }
-        return 0;
-    }
-
-    public lastDrop(id: string, last: number) {
-        this._.users[id] = last;
-        this.save();
-    }
-
-    /**
-     * Validate supply
-     */
-    public hasSupply(date: string, supply: number): boolean {
-        this.fillSupply(date, supply);
-        return this._.supply[date] > 0;
-    }
-
-    public burnSupply(date: string, supply: number) {
-        this.fillSupply(date, supply);
-        this._.supply[date] -= 1;
-        this.save();
-    }
-
-    /**
-     * save status
-     */
-    private save() {
-        fs.writeFileSync(this.path, JSON.stringify(this._, null, 2));
-    }
-
-    private fillSupply(date: string, supply: number) {
-        if (this._.supply[date] === undefined) {
-            this._.supply[date] = supply;
-            this.save();
-        }
-    }
-}
-
-export default BotDb;
